feat: add dev and prod sections for config (#2139)
