+++ next/.gitignore
@@ -28,8 +28,10 @@ yarn-debug.log*
 yarn-error.log*
 .pnpm-debug.log*
 
-# local env files
-.env*.local
+# env files
+/.env*.local
+/.env.development
+/.env.production
 
 # vercel
 .vercel
