+++ next/scripts/build.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+# init
+
+cp ./site-config/$1/prod/.env .env.production
