+++ next/scripts/common-build.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+# init
+
+cp ./site-config/$1/$2/.env .env.production
