+++ next/site-config/data-portal/dev/.env
@@ -1,0 +1,4 @@
+NEXT_PUBLIC_SITE_CONFIG='data-portal-dev'
+NEXT_PUBLIC_SITEMAP_DOMAIN='https://data.humancellatlas.dev.clevercanary.com'
+NEXT_PUBLIC_BASE_PATH=''
+SITEMAP_EXCLUDE=''
