+++ next/site-config/data-portal/dev/constants.ts
@@ -1,0 +1,19 @@
+import {
+  SOCIAL,
+  Social,
+} from "@clevercanary/data-explorer-ui/lib/components/common/Socials/socials";
+
+export const socials: Social[] = [
+  {
+    ...SOCIAL.TWITTER,
+    url: "https://twitter.com/humancellatlas",
+  },
+  {
+    ...SOCIAL.GITHUB,
+    url: "https://github.com/HumanCellAtlas",
+  },
+  {
+    ...SOCIAL.SLACK,
+    url: "https://humancellatlas.slack.com/archives/C02TM2SDVM2",
+  },
+];
