+++ next/site-config/data-portal/prod/.env
@@ -1,0 +1,4 @@
+NEXT_PUBLIC_SITE_CONFIG='data-portal-prod'
+NEXT_PUBLIC_SITEMAP_DOMAIN='https://data.humancellatlas.org'
+NEXT_PUBLIC_BASE_PATH=''
+SITEMAP_EXCLUDE=''
