+++ next/site-config/data-portal/prod/config.ts
@@ -1,0 +1,8 @@
+import { SiteConfig } from "@clevercanary/data-explorer-ui/lib/config/entities";
+import devConfig from "../dev/config";
+
+const config: SiteConfig = {
+  ...devConfig,
+};
+
+export default config;
